+++ Procfile
@@ -1,2 +1,1 @@
-#web: bundle exec rackup config.ru -p $PORT
 slackattend: bundle exec rackup config.ru -p 8765 -o 0.0.0.0
