+++ lib/slackattend.rb
@@ -12,8 +12,9 @@ Encoding.default_external = Encoding.find('UTF-8')
 
 %w(
   core
-  slack_client
   status_log
   current_member
+  slack_client
   backend
+  app
 ).each { |name| require_dependency File.expand_path("../slackattend/#{name}", __FILE__) }
