+++ lib/slackattend/core.rb
@@ -1,9 +1,15 @@
 module Slackattend
   module Core
     def config
-      @config ||= YAML.load_file(File.expand_path('../../../config.yml', __FILE__))
+      @config ||= {}
+    end
+
+    def start
+      YAML.load_file(File.expand_path('../../../config.yml', __FILE__)).each{ |k,v| config[k.to_sym] = v }
+      setup
     end
   end
 
   extend Core
-end 
+end
+
