+++ lib/views/index.haml
@@ -1,9 +1,9 @@
 #in.tile-group.five
   .tile-group-title In-room
-  - @in_rooms.each do |m|
+  - @ins.each do |m|
     .tile.animated.fadeIn{"data-click" => "transform"}
       .tile-content.image
-        %img{:src => "#{m.avatar}", "id" => "#{m.name}", "alt" => "in"}/
+        %img{:src => "#{m.avatar}", "id" => "#{m.name}", "alt" => "#{@in_action}"}/
       .brand
         .label.fg-amber #{m.name}
         .badge.available
@@ -15,5 +15,5 @@
   - @absents.each do |m|
     %a.tile.half.bg-dark.animated.fadeIn{"data-click" => "transform"}
       .tile-content
-        %img{:src => "#{m.avatar}", "id" => "#{m.name}", "alt" => "out"}/
+        %img{:src => "#{m.avatar}", "id" => "#{m.name}", "alt" => "#{@out_action}"}/
 / End tile group
